misc/HtmlTestSuite/results/test-0589.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<title>CSS Ruby Test: Test line breaking before and after &lt;ruby&gt; elements.</title>
<link href="https://drafts.csswg.org/css-ruby-1/#ruby-layout" rel="help"/>
<meta content="est line breaking before and after &lt;ruby&gt; elements" name="assert"/>
<meta content="dom" name="flags"/>
<link href="mailto:kojiishi@gmail.com" rel="author" title="Koji Ishii"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<style>
#container &gt; div {
    line-height:2;
    margin-bottom:1em;
}
</style>
</head><body><div id="log"/>
<div id="container">
    <div style="width:2em;" title="OP+ID">あ（かき</div>
    <div style="width:2em;" title="ID+CL">あい）か</div>
    <div style="width:2em;" title="span+CL">あ<span>す</span>）か</div>
    <div style="width:2em;" title="OP+ruby">あ（<ruby>べ<rt>る</rt></ruby>か</div>
    <div style="width:2em;" title="ruby+CL">あ<ruby>べ<rt>る</rt></ruby>）か</div>
    <div style="width:3em;" title="OP+ruby+CL">あ（<ruby>べ<rt>る</rt></ruby>）か</div>
</div>
<script>
runTests();
function runTests() {
    var lineHeight = parseFloat(getComputedStyle(container.firstElementChild).lineHeight);
    Array.prototype.forEach.call(container.children, function (element) {
        test(function () {
            var lineCount = element.offsetHeight / lineHeight;
            assert_approx_equals(lineCount, 3, 0.1);
        }, element.title);
    });
}
</script>
</body></html>
